a script that calculates monthly savings based on inputted monthly income and expenses update

+++ python_introduction/finance_calculator.py
@@ -1,9 +1,9 @@
-monthly_income = int(input('Enter your monthly income: '))
-monthly_expenses = int(input('Enter your total monthly expenses: '))
+monthly_income = float(input('Enter your monthly income: '))
+monthly_expenses = float(input('Enter your total monthly expenses: '))
 
 monthly_savings = (monthly_income - monthly_expenses)
 
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 
-print('Your monthly savings are $', float(monthly_savings))
-print('Projected savings after one year, with interest, is: $', float(projected_savings))
+print('Your monthly savings are $', monthly_savings)
+print('Projected savings after one year, with interest, is: $', projected_savings)
